pyproject.toml
[project]
name = "smart-manufact"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "pymupdf>=1.26.7",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]
